Add option to flush streams when memory threshold reached
The memory consumption of a job is often not foreseeable,
especially when the amount of data is highly inconsistent.
It would either require to set the max batch size low enough
to accommodate every possible load pattern, or to over-provision
memory.
This change adds a safeguard against OOM events by flushing all
streams as soon as the memory consumption reaches a certain threshold.